Use of LogInManager
